requirements.txt
httpx
fastmcp
python-dotenv
pytest
pytest-asyncio